backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.3
python-dotenv==1.0.1
httpx==0.27.2
requests>=2.31.0
blackboxprotobuf>=1.0.1
protobuf>=3.10.0
